Prepare static linking support
In order to statically link azul as a regular Rust library,
the function calls including arguments and return types have
to 1:1 match each other. This is a WIP commit, so that the
function body of the DLL function just forwards the call to
the library-internal object using UFCS.

Ideally, the library should be compiled as statically-first,
dynamically-second.